Refactor user dto
