--- Shuttle-back/src/main/java/com/shuttle/user/UserDTO.java
@@ -1,70 +1,0 @@
-package com.shuttle.user;
-
-import com.shuttle.common.Entity;
-
-public class UserDTO extends Entity {
-    private String name;
-    private String surname;
-    private String profilePicture;
-    private String telephoneNumber;
-    private String email;
-    private String address;
-
-    public UserDTO(Long id,String name, String surname, String profilePicture, String telephoneNumber, String email, String address) {
-        this.setId(id);
-        this.name = name;
-        this.surname = surname;
-        this.profilePicture = profilePicture;
-        this.telephoneNumber = telephoneNumber;
-        this.email = email;
-        this.address = address;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public void setSurname(String surname) {
-        this.surname = surname;
-    }
-
-    public void setProfilePicture(String profilePicture) {
-        this.profilePicture = profilePicture;
-    }
-
-    public void setTelephoneNumber(String telephoneNumber) {
-        this.telephoneNumber = telephoneNumber;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public void setAddress(String address) {
-        this.address = address;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public String getSurname() {
-        return surname;
-    }
-
-    public String getProfilePicture() {
-        return profilePicture;
-    }
-
-    public String getTelephoneNumber() {
-        return telephoneNumber;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public String getAddress() {
-        return address;
-    }
-}
